Class Checks and CheckIfRawMaterialIsUsed method

+++ proderp/src/main/java/pijavaparty/proderp/GUI/RawMaterials/RawMaterialDelete.java
@@ -5,6 +5,7 @@ import java.awt.Toolkit;
 import javax.swing.JOptionPane;
 import pijavaparty.proderp.dao.RawMaterialDao;
 import pijavaparty.proderp.GUI.LogIn;
+import pijavaparty.proderp.Services.Checks;
 
 /**
  *
@@ -104,10 +105,16 @@ public class RawMaterialDelete extends javax.swing.JFrame {
                 throw new Exception();
             String value1_id=value_id.getText();
             int newvalue1_id=Integer.parseInt(value1_id);
-            RawMaterialDao rawdao=new RawMaterialDao();
-            rawdao.delete(newvalue1_id);
-            JOptionPane.showMessageDialog(null,"Deleted");
-            dispose();
+            Checks check = new Checks();
+            if (check.CheckIfRawMaterialIsUsed(newvalue1_id)) {
+                JOptionPane.showMessageDialog(null,"Cannot be deleted. Raw Material used in a product's recipe.","Error",  JOptionPane.ERROR_MESSAGE);
+            } else {
+                RawMaterialDao rawdao=new RawMaterialDao();
+                rawdao.delete(newvalue1_id);
+                JOptionPane.showMessageDialog(null,"Deleted");
+                dispose();
+            }
+            
         }  catch (Exception e){
             JOptionPane.showMessageDialog(null,"Insert Quantity.","Error",  JOptionPane.ERROR_MESSAGE);
             RawMaterialDelete stor = new RawMaterialDelete();
